+++ rest-test/src/test/java/org/rhq/metrics/rest/BaseTest.java
@@ -32,48 +32,12 @@ import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.Matchers.emptyCollectionOf;
 import static org.hamcrest.Matchers.hasSize;
 
-@RunWith(Arquillian.class)
+
 public class BaseTest extends AbstractTestBase {
 
     private static final String APPLICATION_JAVASCRIPT = "application/javascript";
     private static final long SIXTY_SECONDS = 60*1000L;
 
-    @ArquillianResource
-    private URL baseUrl;
-
-    @Deployment(testable=false)
-    public static WebArchive createDeployment() {
-        File pomFile = new File("../rest-servlet/pom.xml");
-
-        System.out.println("Pom file path: " + pomFile.getAbsolutePath());
-        System.out.flush();
-
-        WebArchive archive =
-        ShrinkWrap.create(MavenImporter.class)
-            .offline()
-            .loadPomFromFile(pomFile)
-            .importBuildOutput()
-            .as(WebArchive.class);
-        System.out.println("archive is " + archive.toString(false));
-        System.out.flush();
-        return archive;
-    }
-
-    @Before
-    public void setupRestAssured() {
-        if (baseUrl!=null) {
-            RestAssured.baseURI = baseUrl.toString();
-        } else {
-            RestAssured.baseURI = "http://localhost:8080/rhq-metrics";
-        }
-        // There is no need to set RestAssured.basePath as this is already in the baseUrl,
-        // set via Arquillian in the baseUrl
-    }
-
-    @After
-    public void tearDown() {
-        RestAssured.reset();
-    }
 
 
     @Test
